% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quads2d.R
\name{quads2d}
\alias{quads2d}
\title{Add Quads to a Plot}
\usage{
quads2d(
  x,
  y = NULL,
  border = NULL,
  col = NA,
  ...,
  alpha = 1,
  lty = 1,
  lwd = 1,
  asp = "",
  add = TRUE
)
}
\arguments{
\item{x, y}{coordinates any reasonable way of defining the coordinates is
acceptable. See the function \code{\link[=xy.coords]{xy.coords()}} for details}

\item{border}{border line colour}

\item{col}{fill colour}

\item{...}{further arguments passed to \code{\link[=grid.polygon]{grid.polygon()}} WIP}

\item{alpha}{alpha transparency}

\item{lty}{line type}

\item{lwd}{line width}

\item{asp}{control the aspect ratio of the plot (only when \code{add = TRUE})}

\item{add}{logical, add to current plot or start a new one}
}
\value{
called for its side effect, drawing quads on a plot
}
\description{
Draw quads with quadruplets of points, quadruplets are defined sequentially.
}
\details{
It is expected that there will be a number of coordinates exactly divisible by 4.

This is a convenience function (analogous to \code{rgl}'s \code{quads3d()}) that
doesn't require reshaping sequential coordinates into a new structure.

If there's no current plot this function will start one, this may not be a good
idea.
}
\examples{
quads2d(runif(36), runif(36), add = FALSE, col = sample(colors(), 9))
}
