% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments2d.R
\name{segments2d}
\alias{segments2d}
\title{Add Line Segments to a Plot}
\usage{
segments2d(x, y = NULL, ..., asp = "", add = TRUE)
}
\arguments{
\item{x, y}{coordinates any reasonable way of defining the coordinates is
acceptable. See the function \code{\link[=xy.coords]{xy.coords()}} for details}

\item{...}{further arguments passed to \code{\link[=segments]{segments()}}}

\item{asp}{control the aspect ratio of the plot (only when \code{add = TRUE})}

\item{add}{logical, add to current plot or start a new one}
}
\value{
called for its side effect, drawing line segments on a plot
}
\description{
Draw line segments between pairs of points, pairs are defined sequentially.
}
\details{
It is expected that there will be an even number of coordinates.

This is a convenience function (analogous to \code{rgl}'s \code{segments3d()}) that
doesn't require reshaping sequential coordinates into start,end pairs.

If there's no current plot this function will start one, this may not be a good
idea.
}
\examples{
segments2d(runif(10), runif(10), add = FALSE)
}
