% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangles2d.R
\name{triangles2d}
\alias{triangles2d}
\title{Add Triangles to a Plot}
\usage{
triangles2d(
  x,
  y = NULL,
  border = NULL,
  col = NA,
  ...,
  alpha = 1,
  lwd = 1,
  lty = 1,
  asp = "",
  add = TRUE
)
}
\arguments{
\item{x, y}{coordinates any reasonable way of defining the coordinates is
acceptable. See the function \code{\link[=xy.coords]{xy.coords()}} for details}

\item{border}{border line colour}

\item{col}{fill colour}

\item{...}{further arguments passed to \code{\link[=grid.polygon]{grid.polygon()}}}

\item{lwd}{line width}

\item{lty}{line type}

\item{asp}{control the aspect ratio of the plot (only when \code{add = TRUE})}

\item{add}{logical, add to current plot or start a new one}
}
\value{
called for its side effect, drawing triangles on a plot
}
\description{
Draw triangles with triplets of points, triplets are defined sequentially.
}
\details{
It is expected that there will be a number of coordinates exactly divisible by 3.

This is a convenience function (analogous to \code{rgl}'s \code{triangles3d()}) that
doesn't require reshaping sequential coordinates into a new structure.

If there's no current plot this function will start one, this may not be a good
idea.

TODO: add all line rendering options for borders
}
\examples{
triangles2d(runif(27), runif(27), add = FALSE, col = sample(colors(), 9))
}
